Improve app stability and user experience with various updates
Includes:
- Thermal throttling check fix using "thermalservice" string.
- UI enhancements with imported icons and updated logic for About, EducationHub, Help, HelpViewer, OverlayButton, and PerformanceDashboard screens.
- Refactored `OnboardingActivity` and `OnboardingFlow` for improved lifecycle management and permission handling.
- Updated `PatternPlanScreen` and `ProofCapsuleScreen` with integer IDs for demo patterns.
- Added `ProUpgradePrompt` and `PaywallScreen` composables for premium feature access.
- Introduced `PatternMatchExtensions` with `boundingBox` and `explanation` properties.
- Converted `confidence` to Float in `EnhancedOverlayRenderer` and `EnhancedOverlayView`.

+++ app/src/main/java/com/lamontlabs/quantravision/ui/EnhancedOverlayRenderer.kt
@@ -81,7 +81,7 @@ class EnhancedOverlayRenderer(
 
     private fun drawEnhancedPattern(canvas: Canvas, match: PatternMatch) {
         val rect = match.boundingBox
-        val confidence = match.confidence
+        val confidence = match.confidence.toFloat()
         val style = PatternStyle.forPattern(match.patternName, confidence)
         
         val alpha = (255 * animationProgress).toInt()
