+++ app/src/main/java/com/lamontlabs/quantravision/ui/EnhancedOverlayView.kt
@@ -230,7 +230,7 @@ class EnhancedOverlayView @JvmOverloads constructor(
 
     private fun drawEnhancedPattern(canvas: Canvas, match: PatternMatch) {
         val rect = match.boundingBox
-        val confidence = match.confidence
+        val confidence = match.confidence.toFloat()
         val style = PatternStyle.forPattern(match.patternName, confidence)
         
         val alpha = (255 * animationProgress).toInt()
